refactor: share normalize_date helper
