+++ wsm/parsing/eslog.py
@@ -18,6 +18,7 @@ import xml.etree.ElementTree as ET
 from typing import List, Dict, Optional, Tuple
 
 import pandas as pd
+from .utils import _normalize_date
 
 # Uvoz pomožnih funkcij iz money.py:
 from wsm.parsing.money import extract_total_amount, validate_invoice as validate_line_values
@@ -36,18 +37,6 @@ def _decimal(el: ET.Element | None) -> Decimal:
     except Exception:
         return Decimal("0")
 
-def _normalize_date(date_str: str) -> str:
-    """Convert ``DD.MM.YYYY`` or ``YYYYMMDD`` into ``YYYY-MM-DD``."""
-    s = date_str.replace(" ", "").replace("\xa0", "")
-    m = re.match(r"(\d{1,2})\.(\d{1,2})\.(\d{4})$", s)
-    if m:
-        d, mth, y = m.groups()
-        return f"{y}-{int(mth):02d}-{int(d):02d}"
-    m = re.match(r"(\d{4})(\d{2})(\d{2})$", s)
-    if m:
-        y, mth, d = m.groups()
-        return f"{y}-{mth}-{d}"
-    return s
 
 # Namespace za ESLOG (če je prisoten)
 NS = {"e": "urn:eslog:2.00"}
